ui scope stack (inspired by imgui)
widgets take a static string label/id. in case of id-conflicts (e.g. a
widget in a for-loop), you can push/pop scopes. behind the scenes, the
static pointers and stack are used as input to a hash function that
generates a "unique id" for the widget (the hash is 64bit so collisions
are very rare)

also, lsl_accept() is used to "parse" keyboard input. it works a bit
like an accept function in a parser; it consumes the character if it
matches (and returns 1), otherwise it leaves the input buffer as it were
(and returns 0). currently used to put do into "add node" mode when
pressing spacebar (WIP)