major atls rewrite
colors are now expressions that can refer to colorscheme colors, and be
influenced by "ctx vars". an expression could be something like
`lerp($selected, @normal_color, @selected_color) + @base_color`. `$` is
used for ctx vars, and `@` is used for colorscheme colors.

layer names in gimp are now such expressions, but stand-alone
expressions can also be added with .progs files (e.g. for background and
text colors, but also auxiliary values which are useful because the
expression language lacks user-defined functions)

ctx vars are meant to be altered at runtime. an example could be a
"$selected" ctx var that you set to 1.0 when drawing things that require
an is-selected color (see the lerp example above). but you could also
throw "current time" into a ctx var to get animated colors.

expressions are primarily for colors, but I've considered using them as
replacement for "metadata" as well (layout-related things like position
offsets, for instance).

expressions are compiled to 16bit byte code which is executed like an
AST (and there's no VM stack; the host stack is used for tmp values)

probably a lot of bugs :) and a few things are missing:
 - select() statements are a bit complicated and haven't been
   implemented yet (it's a vararg statement, where the first argument
   evaluates to an index that chooses which of the remaining args
   becomes the value of the statement)
 - select() is sorta required for ctx vars to be useful, so I'm actually
   not using them anywhere yet (so I've temporary broken connection
   color change for selected connections)