protocol stuff and file drop!
file drop works both in sdl3 and emscripten! (does nothing with the file data
for now)

protocol stuff tries to encode/decode using the same function. is it a bad
idea? I seem to remember trying it once but not why I didn't stick with it. the
varint.h stuff should be leb128.h instead I think.