time travel fix
the problem was not clearing the snapshotcache_offset values
when appropriate. it seems like a source of bugs because I've
added a bunch of needed doc_edit()/ms_edit() calls after test_gig.c
started passing. not sure what I think about it because... a missing
ms_edit() can be hard to test for since there there are many ms_edit()
calls, and all of them must be missing before we start to see subtle
"desync" problems. what can be done to improve this?